Refactor of drag_xy and drag_resize.
Instead of moving the multi-select rect to the top when shown,
we keep it at the back, so that other panels can still be
clicked from within it (doesn't overlay them).
This means that we have to handle drag_xy of selected panels,
to notify all other panels and the multi-select rect.