Modify Figure export script to allow for variable scalebar widths.
Achieved by querying the new scalebar property and passing the width to draw_scalebar_line().
For correct positioning, offset parameters for the scalebar and the label are introduced.

Seems to work for PDF in all scalebar positions and a variety of width/length combinations.

Known Issue: TIFF export not working yet.